requirements.txt
streamlit
chromadb
pandas
google.genai
google.api_core